fix(questionnaire): allow immediate retake when can_retake_after is None or zero
Previously, `can_retake_after=None` blocked retakes entirely in the
submission service, inconsistent with the eligibility gate (gates.py)
which already treated None as "immediate retake." Now both paths are
aligned: `max_attempts` alone controls whether retakes are possible,
and `can_retake_after` only controls the cooldown duration.

Also extracts shared eq_* test fixtures into conftest.py and splits
admission resubmission tests into their own file to stay under the
1000-line limit.
